requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
psycopg2==2.7.7
PyJWT==1.7.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
whitenoise==3.3.1
sqlparse==0.3.0
Markdown==3.1
autopep8==1.4.4
bleach==3.1.0
django-markdownify==0.8.0
pycodestyle==2.5.0
six==1.12.0
webencodings==0.5.1
